data_transmission/requirements.txt
# dependencies
pbr>=0.5.21
pam>=0.1.4
WebOb>=1.2.3,<1.3
eventlet>=0.13.0
greenlet>=0.3.2
PasteDeploy>=1.5.0
Paste
Routes>=1.12.3,!=2.0
passlib
iso8601>=0.1.8
Babel>=1.3
itsdangerous
msgpack>=0.4.0
pytz>2013.6
pyparsing>=2.0.7
netaddr>=0.7.13,!=0.7.16
oslo.log>=3.19.0
oslo.utils>=3.18.0
oslo.serialization
oslo.config>=1.2.0
pika>=0.10.0
config>=0.3.9
chardet>=3.0.4
schedule
SQLAlchemy
python-memcached
kazoo
python-logstash
psutil
kafka-python
elasticsearch
uuid
datetime
requests